Added creditCard functiones in db

+++ bikeProject/src/main/java/bikeProject/database/CreditCardDatabaseInterface.java
@@ -5,5 +5,5 @@ import java.sql.SQLException;
 
 public interface CreditCardDatabaseInterface {
 
-    int registerNewCreditCard(long userid, long number, long cvv, Date expireDate) throws SQLException;
+    long registerNewCreditCard(long userid, long number, long cvv, Date expireDate) throws SQLException;
 }
